removed older scripts
